+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 1.0.0-alpha.8: UNRELEASED
 
 - BUGFIX: Remove un-used font files from distribution
+- Render sassdoc comments that aren't attached to a Sass block as Markdown
+  prose.
+- Preserve the original order of sassdoc comments rather than grouping by type.
 
 ## 1.0.0-alpha.7: 2017-06-16